🧪🐛fix (project) Do not recreate `User` and `Worker`
This causes the test to create 2 sets of `User` and `Worker`. They are already being created in the
super class `ClientTestCase`. Another weird bug caused by this is that the `self.user`/`self.worker`
instance used to actaully make the api requests are different from the ones being used in the test
asserts. So tests that require assertions based on `User` or `Worker` will always fail