🐛fix (task) Set parent when creating a group
The issue with not setting the parent is that, this tries to update the group if it exists and to do
that the condition to fetch the group from db is `job=job` but that job itself has job pointing to
itself and jobs are saved in the same db table (task_group). By adding parent we make sure the
correct object is returned. Wihtout parent the job will be included as well